Improved Disassembler
Disassembles and shows upcoming instructions to be executed, next do previous instructions in the machine code. It currently shows what ever instruction is at address zero in memory where I plan to have the previous instructions displayed due to a function not being filled and just returning zero for the time being.