Preparations for large rewrite
I want to improve the accuracy of some of the components (meaning mostly a complete rewrite of most of them), also looking back at this after not touching it for a while, I'm just not satisfied with how I did some things and feel it could be structured better. Having components poking into the memory array to know the values of their IO registers just feels wrong.